Remove truffle from tokenTransferProxy tests
