+++ packages/contracts/test/zrx_token.ts
@@ -3,9 +3,9 @@ import { BigNumber } from '@0xproject/utils';
 import * as chai from 'chai';
 import Web3 = require('web3');
 
-import { Artifacts } from '../../util/artifacts';
-import { constants } from '../../util/constants';
-import { ContractInstance } from '../../util/types';
+import { Artifacts } from '../util/artifacts';
+import { constants } from '../util/constants';
+import { ContractInstance } from '../util/types';
 
 import { chaiSetup } from './utils/chai_setup';
